+++ tcmtool/svn/tsm/src/tcmalg/openssl/crypto.h
@@ -83,20 +83,20 @@ typedef struct crypto_ex_data_st
 
 /* Called when a new object is created */
 typedef int CRYPTO_EX_new(void *parent, void *ptr, CRYPTO_EX_DATA *ad,
-					int idx, long argl, void *argp);
+					int idx, int argl, void *argp);
 /* Called when an object is free()ed */
 typedef void CRYPTO_EX_free(void *parent, void *ptr, CRYPTO_EX_DATA *ad,
-					int idx, long argl, void *argp);
+					int idx, int argl, void *argp);
 /* Called when we need to dup an object */
 typedef int CRYPTO_EX_dup(CRYPTO_EX_DATA *to, CRYPTO_EX_DATA *from, void *from_d, 
-					int idx, long argl, void *argp);
+					int idx, int argl, void *argp);
 
 /* This stuff is basically class callback functions
  * The current classes are SSL_CTX, SSL, SSL_SESSION, and a few more */
 
 typedef struct crypto_ex_data_func_st
 	{
-	long argl;	/* Arbitary long */
+	int argl;	/* Arbitary long */
 	void *argp;	/* Arbitary void * */
 	CRYPTO_EX_new *new_func;
 	CRYPTO_EX_free *free_func;
